running kokkos_hip_stream_ver2.sh only if on APU

+++ tests/kokkos_hip_stream_ver2.sh
@@ -1,18 +1,24 @@
 #!/bin/bash
 
-module load amdclang
-module load rocm
-module load kokkos
+XNACK_COUNT=`rocminfo | grep xnack | wc -l`
+if [ ${XNACK_COUNT} -lt 1 ]; then
+   echo "Skip"
+else
 
-git clone --recursive https://github.com/EssentialsOfParallelComputing/Chapter13 Chapter13
-pushd Chapter13/Kokkos/StreamTriad/Ver2
-sed -i -e 's/80000000/100000/' StreamTriad.cc
+   module load amdclang
+   module load rocm
+   module load kokkos
 
-rm -rf build
-mkdir build && cd build
-CXX=hipcc cmake ..
-make
-./StreamTriad
+   git clone --recursive https://github.com/EssentialsOfParallelComputing/Chapter13 Chapter13
+   pushd Chapter13/Kokkos/StreamTriad/Ver2
+   sed -i -e 's/80000000/100000/' StreamTriad.cc
 
-popd
-rm -rf Chapter13
+   rm -rf build
+   mkdir build && cd build
+   CXX=hipcc cmake ..
+   make
+   ./StreamTriad
+
+   popd
+   rm -rf Chapter13
+fi
